% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computo_article.R
\name{article}
\alias{article}
\title{Template for Computo articles}
\usage{
article(...)
}
\arguments{
\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}
}
\description{
A template for creating scientific articles according to the Computo format, native to the web.
}
\details{
Based on distill articles, which feature attractive, reader-friendly typography, flexible
layout options for visualizations, and full support for footnotes and citations.
}
