% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computo_website.R
\name{computo_website}
\alias{computo_website}
\title{R Markdown site generator for Computo website (A Distill website)}
\usage{
computo_website(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{encoding}{Ignored. The encoding is always assumed to be UTF-8.}

\item{...}{Currently unused.}
}
\description{
R Markdown site generator for Computo website (A Distill website)
}
