% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R, R/pdf.R
\name{html_article}
\alias{html_article}
\alias{pdf_article}
\title{Template for articles (HTML)}
\usage{
html_article(...)

pdf_article(...)
}
\arguments{
\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}
}
\description{
A template for creating HTML articles according to the our futur format.

A template for creating HTML articles according to the our futur format.
}
