% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computo_rticle_pdf.R
\name{pdf_article}
\alias{pdf_article}
\title{Template for articles (HTML)}
\usage{
pdf_article(...)
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}
}
\description{
A template for creating HTML articles according to the our futur format.
}
